jobfit-server/requirements.txt
fastapi
uvicorn
anthropic
python-dotenv
pdfplumber